Compability with Irrlicht 1.3.1 for Windows.
git-svn-id: https://irrlichtnetcp.svn.sourceforge.net/svnroot/irrlichtnetcp@105 2f455fab-421c-0410-8db0-c3f11262bd97

+++ IrrlichtW/device.cpp
@@ -45,7 +45,7 @@ IntPtr CreateDevice(E_DRIVER_TYPE type, M_DIM2DS dim, int bits, bool full, bool 
     return createDeviceEx(params);
 }
 
-IntPtr CreateDeviceA(E_DRIVER_TYPE type, M_DIM2DS dim, int bits, bool full, bool stencil, bool vsync, bool antialias, int handle)
+IntPtr CreateDeviceA(E_DRIVER_TYPE type, M_DIM2DS dim, int bits, bool full, bool stencil, bool vsync, bool antialias, IntPtr handle)
 {
     SIrrlichtCreationParameters params;
     params.AntiAlias = antialias;
